% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heatmap.R
\name{bray_curtis_dist}
\alias{bray_curtis_dist}
\title{Calculate Bray-Curtis distance}
\usage{
bray_curtis_dist(mat)
}
\arguments{
\item{mat}{Numeric matrix (rows are observations)}
}
\value{
A dist object
}
\description{
Bray-Curtis dissimilarity: sum(|xi - yi|) / sum(xi + yi)
}
\keyword{internal}
