% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dendro.R
\name{calculate_assembly_distance}
\alias{calculate_assembly_distance}
\title{Calculate distance between assemblies}
\usage{
calculate_assembly_distance(mat, method)
}
\arguments{
\item{mat}{Matrix with assemblies as rows, orthogroups as columns}

\item{method}{Distance method}
}
\value{
dist object
}
\description{
Calculate distance between assemblies
}
\keyword{internal}
