% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heatmap.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate distance matrix using specified method}
\usage{
calculate_distance(mat, method)
}
\arguments{
\item{mat}{Numeric matrix (rows are observations)}

\item{method}{One of "binary", "jaccard", "bray-curtis"}
}
\value{
A dist object
}
\description{
Calculate distance matrix using specified method
}
\keyword{internal}
