% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dendro.R
\name{dendro_to_data}
\alias{dendro_to_data}
\title{Convert dendrogram to data frames for ggplot2}
\usage{
dendro_to_data(dend)
}
\arguments{
\item{dend}{A dendrogram object}
}
\value{
List with segments and labels data frames
}
\description{
Convert dendrogram to data frames for ggplot2
}
\keyword{internal}
