% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-scores.R
\name{extract_scores}
\alias{extract_scores}
\title{Extract scores from protein collection}
\usage{
extract_scores(proteins, score_column)
}
\arguments{
\item{proteins}{A protein_collection object}

\item{score_column}{Name of the score column}
}
\value{
Data frame with columns: assembly, score
}
\description{
Extract scores from protein collection
}
\keyword{internal}
