% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singletons.R
\name{get_singletons}
\alias{get_singletons}
\title{Get singletons from clustering result}
\usage{
get_singletons(orthogroups)
}
\arguments{
\item{orthogroups}{An \code{orthogroup_result} object.}
}
\value{
A tibble with columns \code{protein_id} and \code{assembly}.
}
\description{
Extracts the singletons tibble from an orthogroup_result object.
}
\examples{
\dontrun{
singletons <- get_singletons(clusters)
}
}
