% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heatmap.R
\name{jaccard_dist}
\alias{jaccard_dist}
\title{Calculate Jaccard distance}
\usage{
jaccard_dist(mat)
}
\arguments{
\item{mat}{Numeric matrix (rows are observations)}
}
\value{
A dist object
}
\description{
Jaccard dissimilarity: 1 - |A ∩ B| / |A ∪ B|
For binary data: 1 - (number of shared 1s) / (number where either is 1)
}
\keyword{internal}
