% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singletons.R
\name{n_singletons}
\alias{n_singletons}
\title{Count singletons}
\usage{
n_singletons(orthogroups)
}
\arguments{
\item{orthogroups}{An \code{orthogroup_result} object.}
}
\value{
Integer count of singletons.
}
\description{
Returns the number of singleton proteins in a clustering result.
}
\examples{
\dontrun{
n <- n_singletons(clusters)
}
}
