% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_pa_matrix}
\alias{new_pa_matrix}
\title{Create a pa_matrix object}
\usage{
new_pa_matrix(
  matrix,
  orthogroups,
  assemblies,
  type = "binary",
  threshold = NULL
)
}
\arguments{
\item{matrix}{Matrix. Rows = orthogroups, cols = assemblies (0/1 or scores).}

\item{orthogroups}{A tibble. Orthogroup metadata.}

\item{assemblies}{A tibble. Assembly metadata.}

\item{type}{Character. One of "binary", "score", or "count".}

\item{threshold}{Numeric. Score threshold used (if applicable).}
}
\value{
A \code{pa_matrix} S3 object.
}
\description{
Presence/absence matrix with metadata.
}
