% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_protein_collection}
\alias{new_protein_collection}
\title{Create a protein_collection object}
\usage{
new_protein_collection(assemblies)
}
\arguments{
\item{assemblies}{List of protein_set objects.}
}
\value{
A \code{protein_collection} S3 object.
}
\description{
Container for multiple assemblies.
}
