% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pan.R
\name{plot-pan}
\alias{plot-pan}
\title{Pan-Genome Structure Visualization}
\description{
Functions for visualizing the structure of the pan-genome (or pan-effectorome),
showing the distribution of proteins across core, accessory, rare, unique,
and singleton categories.
}
\keyword{internal}
