% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{run_mmseqs2}
\alias{run_mmseqs2}
\title{Run MMseqs2 clustering}
\usage{
run_mmseqs2(
  proteins,
  mode = "fast",
  min_identity = 70,
  min_coverage = 50,
  threads = NULL,
  tool_path = NULL,
  conda_env = NULL
)
}
\arguments{
\item{proteins}{A \code{protein_collection} object.}

\item{mode}{Character. "fast" or "thorough".}

\item{min_identity}{Numeric. Minimum sequence identity percent.}

\item{min_coverage}{Numeric. Minimum query coverage percent.}

\item{threads}{Integer. Number of CPU threads.}

\item{tool_path}{Character. Optional explicit path to diamond.}

\item{conda_env}{Character. Optional conda environment.}
}
\value{
An \code{orthogroup_result} object.
}
\description{
Run MMseqs2 clustering
}
\keyword{internal}
