% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{run_orthofinder}
\alias{run_orthofinder}
\title{Run OrthoFinder clustering}
\usage{
run_orthofinder(
  proteins,
  mode = "fast",
  threads = NULL,
  tool_path = NULL,
  conda_env = NULL
)
}
\arguments{
\item{proteins}{A \code{protein_collection} object.}

\item{mode}{Character. "fast" or "thorough".}

\item{threads}{Integer. Number of CPU threads.}

\item{tool_path}{Character. Optional explicit path to diamond.}

\item{conda_env}{Character. Optional conda environment.}
}
\value{
An \code{orthogroup_result} object.
}
\description{
Run OrthoFinder clustering
}
\keyword{internal}
