% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singletons.R
\name{singleton-functions}
\alias{singleton-functions}
\title{Singleton Utility Functions}
\description{
Functions for working with singleton proteins (proteins not assigned to any
orthogroup during clustering).
}
\keyword{internal}
