% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paneffectR-package.R
\docType{package}
\name{paneffectR-package}
\alias{paneffectR}
\alias{paneffectR-package}
\title{paneffectR: Comparative Genomics of Effector Proteins Across Genome Assemblies}
\description{
Clusters proteins from multiple genome assemblies into ortholog groups, builds presence/absence matrices, and generates publication-ready visualizations. Designed for effector protein analysis from the omnieff pipeline but works with any protein sets.
}
\author{
\strong{Maintainer}: Dan MacLean \email{dan.maclean@tsl.ac.uk}

}
\keyword{internal}
